requirements_test.txt
# Requirements for running tests on different python versions
coverage==4.4.1
mock>=1.0.1
flake8>=2.1.0
tox>=1.7.0
codecov>=2.0.0
django-model-utils>=2.0

